+++ src/ProductSchemaServiceProvider.php
@@ -38,10 +38,15 @@ class ProductSchemaServiceProvider extends ServiceProvider
         $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
 
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'lunar-product-schemas');
+        $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'lunar-product-schemas');
 
         $this->publishes([
             __DIR__.'/../config/lunar-product-schemas.php' => config_path('lunar-product-schemas.php'),
         ], 'lunar-product-schemas-config');
+
+        $this->publishes([
+            __DIR__.'/../resources/lang' => $this->app->langPath('vendor/lunar-product-schemas'),
+        ], 'lunar-product-schemas-translations');
 
         if ($this->app->runningInConsole()) {
             $this->commands([